dvide main,model
